# Improve performance for small inputs